requirements.txt
argon2-cffi==20.1.0
crispy-bootstrap5==0.4
Django>=3.2
django-allauth==0.45.0
dj-database-url==0.5.0
django-configurations==2.2
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-registration==3.2
psycopg2>=2.8
djangorestframework
drf-yasg
django-filter